pyproject.toml
[tool.poetry]
name = "nrlf-api"
version = "1.0.0"
description = ""
authors = ["NHS Digital"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
aws-lambda-powertools = "^2.19.0"
nhs-number = "==1.3.4"
pydantic = "==2.6.3"
pytest = "^7.4.0"
boto3 = "^1.34.15"
pyjwt = "^2.8.0"
cryptography = "^41.0.3"
more-itertools = "^10.1.0"
typing-extensions = "^4.7.1"
requests = "^2.31.0"
jsonschema = {version = "4.17.*", extras = ["format"]}
nrlf-lambda-pipeline = {url = "https://github.com/NHSDigital/nrlf-lambda-pipeline/releases/download/v0.6.0/nrlf_lambda_pipeline-0.3.0-py3-none-any.whl"}
pydantic-settings = "^2.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
sh = "^2.0.4"
awscli = "^1.29.30"
moto = "^4.1.13"
behave = "^1.2.6"
datamodel-code-generator = "^0.21.4"
pyyaml = "^6.0"
pre-commit = "^3.3.3"
flake8 = "^6.0.0"
flake8-print = "^5.0.0"
black = "^23.7.0"
pyhcl = "^0.4.4"
requests = "^2.28.1"
build-scripts = { path = "./scripts/build_scripts/", develop = true }
nrlf = { path = "./layer/nrlf/", develop = true }
lambda-utils = { path = "./layer/lambda_utils/", develop = true }
helpers = { path = "./helpers/", develop = true }
fire = "^0.5.0"
gherkin-official = "^24.1.0"
ez-setup = "^0.9"
pygments = "^2.14.0"
hypothesis = "^6.82.6"
nrlf-converter = {url = "https://github.com/NHSDigital/nrlf-converter/releases/download/0.0.9/nrlf_converter-0.0.9-py3-none-any.whl"}
allure-pytest = "^2.13.2"
allure-behave = "^2.13.2"
urllib3 = "^1.26.18"
boto3-stubs = {version = "1.34.54", extras = ["dynamodb", "ecr", "ecs", "firehose", "lambda", "rds", "s3", "secretsmanager", "sts"]}

[tool.poetry.group.legacy]
optional = true

[tool.poetry.group.legacy.dependencies]
libffx = "0.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.sqlfluff.core]
dialect = "postgres"

[tool.sqlfluff.rules.capitalisation.identifiers]
capitalisation_policy = "consistent"
